UPDATE: organize axi bundle definitions
